backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain-openai==0.1.0
langchain-community==0.0.20
langchain-text-splitters==0.0.1
langchain-chroma==0.1.0
langchain-core==0.1.0
openai==1.12.0
chromadb==0.4.22
unstructured==0.11.8
pypdf==4.0.1

